+++ src/pyopticfilm/device/__init__.py
@@ -6,6 +6,7 @@ from pyopticfilm.device.model_7200i import MODEL_7200_V2, MODEL_7200I, Model7200
 from pyopticfilm.device.model_7300 import MODEL_7300, MODEL_7400_V1, Model7300
 from pyopticfilm.device.model_7400 import MODEL_7400, MODEL_8100, Model7400
 from pyopticfilm.device.model_7500i import MODEL_7500I, MODEL_7600I_V1, Model7500i
+from pyopticfilm.device.model_8100_v2 import MODEL_8100_V2, Model8100V2
 from pyopticfilm.device.model_8200i import MODEL_8200I, Model8200i
 from pyopticfilm.device.model_8200i_se import MODEL_8200I_SE, Model8200iSE
 from pyopticfilm.device.protocol import AsicDriver, FilmModel, MotorProfile
@@ -37,6 +38,7 @@ __all__ = [
     "MODEL_7600I_V1",
     "MODEL_7600I_V2",
     "MODEL_8100",
+    "MODEL_8100_V2",
     "MODEL_8200I",
     "MODEL_8200I_SE",
     "AsicDriver",
@@ -46,6 +48,7 @@ __all__ = [
     "Model7300",
     "Model7400",
     "Model7500i",
+    "Model8100V2",
     "Model8200i",
     "Model8200iSE",
     "MotorProfile",